initial refactoring
1. rm Emitter2.jl
2. split Emitters.jl into separate submodule files
3. mv Emitters.Visualization -> Vis
4. rm hacky root-level include("Emitters.jl")
5. fix incorrect import statements
