+++ src/OpticSim.jl
@@ -40,12 +40,7 @@ include("constants.jl")
 include("utilities.jl")
 include("Geometry/Geometry.jl")
 include("Optical/Optical.jl")
-include("Visualization/Visualization.jl")
-
-# TODO: RG: the following include is seperated from the Optical include due to the new approach of putting the visualazation
-# part with the logic. We need to come up with a better approch to how and where visualization code should reside.
-include("Optical/Emitters.jl")      # defines the Emitters module
-
+include("Vis/Vis.jl")
 include("Examples/Examples.jl")
 include("Optimization/Optimization.jl")
